Fix move up/down actually failing to persist (missing numeric cast)
Root cause of the recurring "move up/down does it, then reverts to old"
report: the item-order swap's UPDATE used a bare CASE WHEN ... THEN \$2
... THEN \$4 END with no type hint, and Postgres could not infer a
numeric type for it against the numeric item_order column, throwing
"column is of type numeric but expression is of type text" -- so every
move request 500'd. The frontend's optimistic UI showed the swap
immediately, then the failed request's catch block refetched the
(unchanged) draft, which is exactly the "reverts to old" symptom. The
equivalent node-reorder query already had \$2::numeric/\$4::numeric
casts; this one just never got them. Reproduced the exact error against
the dev DB with the unfixed query, confirmed the cast resolves it, and
verified two consecutive moves land in the correct final order.

+++ backend/src/routes/builder.js
@@ -598,7 +598,7 @@ router.post("/items/:itemId/move", async (req, res, next) => {
     const neighbor = neighborResult.rows[0];
     if (!neighbor) return respondWithDraft(req, res, req.user, item.plan);
     await query(
-      `update plans.plan_items set item_order = case when id = $1 then $2 when id = $3 then $4 end, updated_at = now() where id in ($1, $3)`,
+      `update plans.plan_items set item_order = case when id = $1 then $2::numeric when id = $3 then $4::numeric end, updated_at = now() where id in ($1, $3)`,
       [item.id, neighbor.item_order, neighbor.id, item.item_order],
     );
     return respondWithDraft(req, res, req.user, item.plan);
